XBOX controls (not yet tested), weapon class
The weapon class is fully operational. Projectiles fire and move
correctly (still needs to be tested on XBOX, where movement is not so
limited). Note: Projectiles disappear when they are no longer active but
they do stay in the collection, this is a problem we need to address
when we get to optimization.

Let me know if I missed anything. Will have a look at NPC's targeting
system.